requirements.txt
numpy>=1.21
opencv-python>=4.5
mediapipe>=0.10
tensorflow>=2.11
scikit-learn>=1.0
joblib>=1.1
cvzone>=1.5
pyttsx3>=2.90
prompt_toolkit>=3.0
matplotlib>=3.5
Pillow>=9.0
